requirements.txt
clipboard==0.0.4
lxml==4.7.1
PyAutoGUI==0.9.53
pyotp==2.6.0
pyperclip==1.8.2
python-dotenv==0.19.2
Pillow==9.0.1
opencv-python